try and improve model code related to accessing entity properties
