+++ libs/@local/hash-backend-utils/package.json
@@ -25,6 +25,7 @@
     "@blockprotocol/core": "0.1.3",
     "@blockprotocol/type-system": "0.1.1",
     "@linear/sdk": "6.0.0",
+    "@local/advanced-types": "0.0.0-private",
     "@local/hash-graph-client": "0.0.0-private",
     "@local/hash-isomorphic-utils": "0.0.0-private",
     "@local/hash-subgraph": "0.0.0-private",
@@ -57,3 +58,4 @@
     "vitest": "1.4.0"
   }
 }
+
